feat(devboard): Add support for ESP32 (#39)

+++ src/lib/CRSFforArduino/src/SerialReceiver/CRSF/CRSF.cpp
@@ -100,7 +100,7 @@ namespace serialReceiver
         }
 
         // Assume the full frame lenthg is 5 bytes until the frame length byte is received.
-        const int fullFrameLength = framePosition < 3 ? 5 : min(rxFrame.frame.frameLength + CRSF_FRAME_LENGTH_ADDRESS + CRSF_FRAME_LENGTH_FRAMELENGTH, CRSF_FRAME_SIZE_MAX);
+        const int fullFrameLength = framePosition < 3 ? 5 : min(rxFrame.frame.frameLength + CRSF_FRAME_LENGTH_ADDRESS + CRSF_FRAME_LENGTH_FRAMELENGTH, (int)CRSF_FRAME_SIZE_MAX);
 
         if (framePosition < fullFrameLength)
         {
